[task]1.闪屏页（冷启动）
