+++ app/src/main/java/com/lz/ZMApplication.java
@@ -1,5 +1,7 @@
 package com.lz;
 
+import android.content.Context;
+
 import com.zm.lib.LibApplication;
 
 /**
@@ -8,5 +10,15 @@ import com.zm.lib.LibApplication;
 
 public class ZMApplication extends LibApplication {
 
+    private static  Context context = null;
 
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        context = getApplicationContext();
+    }
+
+    public static Context getContext() {
+        return context;
+    }
 }
