+++ app/src/main/java/com/lz/module/main/MainActivity.java
@@ -1,8 +1,12 @@
 package com.lz.module.main;
 
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.widget.TextView;
 
+import com.lz.BuildConfig;
 import com.lz.R;
 
 public class MainActivity extends AppCompatActivity {
@@ -11,5 +15,12 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        ((TextView)findViewById(R.id.tv_content)).setText(BuildConfig.API_HOST);
+    }
+
+    public static void startActivity(Context context) {
+        Intent intent = new Intent(context, MainActivity.class);
+        context.startActivity(intent);
     }
 }
