+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,7 @@
     <color name="colorPrimary">#3F51B5</color>
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
+
+    <!-- 通用色板 -->
+    <color name="read">#cc0000</color>
 </resources>
