+++ config.gradle
@@ -4,7 +4,7 @@
 
 ext {
     android = [
-            APPLICATIONID               : 'com.zm',
+            APPLICATIONID               : 'com.lz',
             ANDROID_MIN_SDH_VERSION     : 15,
             ANDROID_TARGET_SDK_VERSION  : 25,
             ANDROID_COMPILE_SDK_VERSION : 25,
